challenge1/dbt/requirements.txt
agate==1.6.1
asn1crypto==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.19.0
azure-storage-blob==12.9.0
Babel==2.9.1
boto3==1.19.3
botocore==1.22.3
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==3.4.8
dbt==0.21.0
dbt-bigquery==0.21.0
dbt-core==0.21.0
dbt-extractor==0.4.0
dbt-postgres==0.21.0
dbt-redshift==0.21.0
dbt-snowflake==0.21.0
google-api-core==1.31.3
google-auth==1.35.0
google-cloud-bigquery==2.28.1
google-cloud-core==1.7.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
hologram==0.0.14
idna==3.3
importlib-metadata==4.8.1
isodate==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
json-rpc==1.13.0
jsonschema==3.1.1
keyring==21.8.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.5
minimal-snowplow-tracker==0.0.2
msgpack==1.0.2
msrest==0.6.21
networkx==2.6.3
oauthlib==3.1.1
oscrypto==1.2.1
packaging==20.9
parsedatetime==2.6
proto-plus==1.19.6
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.11.0
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==3.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
snowflake-connector-python==2.5.1
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
